requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0

# API framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Data processing
pandas>=1.5.0
numpy>=1.24.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Logging and monitoring
loguru>=0.7.0

# Development tools
pytest>=7.0.0
requests>=2.28.0